[TM] Move types required by the spec inside the same file.
This is required for now.
